+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ New features
 * Add modeling.calculate_inf_from_corrse
 * Add modeling.calculate_corr_from_inf
 * Add modeling.create_report
+* Add modeling.check_high_correlations
 
 0.48.0 (2022-02-04)
 -------------------